Suggested Framework/UI Design
The suggested version includes a redesigned UI with the following changes:

-Imported external fonts and adjusted opacity of the calculator UI
-Created Helper class for incorporating static methods which will help to lessen code redundancy.
-Created the SnapShots JPanel which was substituted to the default SubPanel to incorporate the SnapShot UI within the program toggable space.
-Edited the UI design(NOTE: the columns on SnapShots class are made of JPanels)

+++ CrazyCalculator/src/main/Main.java
@@ -20,7 +20,7 @@ public class Main
     gui.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
     gui.setUndecorated(true);
     gui.setLocationRelativeTo(null);
-    gui.setBackground(new Color(0, 0, 0, 0));
+    gui.setBackground(new Color((float)0, (float)0, (float)0, (float)0.00));
     gui.setVisible(true);
   }
